requirements.txt
Django==3.2.10
psycopg2>=2.8.6,<2.9
uWSGI>=2.0.19.1,<2.1
djangorestframework==3.11.0